Разделение простыни кода на файлы-классы
