+++ TurretGenerator.py
@@ -1,0 +1,19 @@
+from SpriteGroup import SpriteGroup
+from Turret import Turret
+
+class TurretGenerator:
+
+    def __init__(self, game_controller):
+        self.turrets = SpriteGroup()
+        self.game_controller = game_controller
+
+    def generate_turrets(self, level):
+        levels = self.game_controller.levels
+        textures = self.game_controller.textures
+        turret_list = levels["turrets"][str(level)]
+        pos = levels["turret_positions"]
+        for i in range(len(turret_list)):
+            self.turrets.add(Turret(self.turrets, pos[str(i + 1)], (72, 72),
+                                    image=textures[turret_list[i]],
+                                    game_controller=self.game_controller,
+                                    turret_type=turret_list[i]))
