+++ constants.py
@@ -1,0 +1,1 @@
+FPS = 25
