+++ levels.json
@@ -10,8 +10,8 @@
   },
   "turret_positions":
   {
-    "1": [50, 100],
-    "2": [100, 200],
-    "3": [150, 300]
+    "1": [100, 300],
+    "2": [150, 400],
+    "3": [200, 500]
   }
 }
